Move utils tests & update
